Brings the branch up to current `main` — 105 commits, including the legacy
tree deletion, the `docs/` reorganisation, the latency-capture rework, the
withdrawal of the `Signal` facade, the six new combinators, and #782 (an
`#[op]` impl now expands from any crate through a generated extension trait).

Four files conflicted textually and two more needed real reconciliation.

**`NodeInfo` collided head-on, and the two are the same thing.** `main` landed
introspection while this branch was out, putting a public `NodeInfo` in
`introspect.rs` — index, label, activation, edges, `removed` — projected out of
`NodeRt` by `interp::node_info`. This branch had added its own `NodeInfo` in
`interp.rs` for `describe()`, with the same five fields plus the wiring record
codegen needs (`src`, `loc`, `build`, `cfg_src`, `takes_closure_cfg`,
`passive_mask`, `has_init_arg`, `unresolved_captures`). Two structs of the same
name in one crate, both projected from `NodeRt` a few hundred lines apart, is
not a merge artifact worth preserving: they are unified into the `introspect`
one, and both `describe()` and `snapshot()` now go through the single
`node_info` projection. `Runner::describe` hands the values over bare;
`snapshot()` wraps them in a `GraphSnapshot` with the renderers.

Three consequences, each deliberate:

- The wiring-record fields are owned (`String`, not `&'static str`), because
  the type derives `Deserialize` and a borrowed field cannot. `describe()` is
  wiring-time, not an execution path, so the allocations cost nothing that
  matters. `codegen::Ineligible` follows for the same reason.
- `GraphSnapshot::to_json` is hand-written specifically so it can be diffed
  against what serde derives (`json_round_trips_through_serde` asserts byte
  identity), so it grew the eight new fields and three helpers to render them
  the way serde does. The new fields are `#[serde(default)]`, so JSON written
  before this still parses.
- `Builder::set_node_build` is now `pub`. Since #782 a `#[op]` impl expands in
  any crate, so every seam the generated builder calls has to be reachable from
  outside this one — `push_node` and `set_node_cfg_src` already were.

**`must_use_combinators.stderr` regenerated.** `#[track_caller]` on the
generated fluent methods makes rustc name them by def path in the
`unused_must_use` diagnostic (`wingfoil::fluent::StreamOps::filter_value`
rather than `filter_value`) for the seven that were not already qualified. All
thirteen lints still fire; only the printed path moved.

The rest were ordinary:

- `ops.rs` — took `main`'s catalog (six new combinators) and re-applied
  `emit_cfg` to the six concrete-config ops. `Limit::Cfg` is `usize` on `main`,
  so `auto_cfg` pins `100usize`.
- `wingfoil-derive` — `cfg_is_closure` had taken `expand_builder`'s doc comment
  when it was inserted; it now carries its own and sits above it, and
  `expand_builder` keeps `main`'s extension-trait section. Dropped the
  `expand_signal` call with the facade it wired.
- `deviation-register` — `main`'s C4 wording, with this branch's two additions
  grafted in (busy-poll ingest generates; one `channel` wiring serves all three
  run shapes). This branch's D28 renumbered to D30: `main`'s D28 and D29 are
  referenced by five other files.
- `examples/core/README.md` — `main` consolidated four Nitro examples into
  `dual_mode`; the `codegen` row rejoins it.
- Stale paths: `docs/wired-graph-codegen-decision.md` is
  `docs/planning/proposals/wired-graph-codegen.md` since the reorg, and
  `docs/deviation-register.md` is under `docs/planning/`.

Verified: `cargo fmt --all --check`, `cargo lint`, `cargo lint-all`,
`cargo test -p wingfoil --all-features` (87 test binaries green; the 9 that
fail all end in `_integration` and want a Docker daemon this sandbox has not
got — the set CI excludes), `cargo test -p wingfoil-derive`, doctests,
`check-example-docs.sh`, `check-python-bindings.sh`, and the `codegen` example
runs with its checked-in artifacts still current.

+++ crates/wingfoil/tests/spawn.rs
@@ -20,10 +20,10 @@ use wingfoil::{NanoTime, RunFor, RunMode};
 #[test]
 fn spawn_source_replays_deterministically_in_historical_mode() {
     let period = Duration::from_millis(10);
-    let n = 6u64;
+    let n = 6usize;
 
     let g = GraphBuilder::new();
-    let counts: Stream<Burst<u64>> = g.spawn(move |wg| wg.ticker(period).count().limit(n as u32));
+    let counts: Stream<Burst<u64>> = g.spawn(move |wg| wg.ticker(period).count().limit(n));
     let scaled = counts
         .map(|b: &Burst<u64>| b.iter().map(|x| x * 10).collect::<Vec<u64>>())
         .with_time();
@@ -58,10 +58,10 @@ fn spawn_source_replays_deterministically_in_historical_mode() {
 #[test]
 fn spawn_source_delivers_all_values_in_realtime() {
     let period = Duration::from_millis(2);
-    let n = 6u64;
+    let n = 6usize;
 
     let g = GraphBuilder::new();
-    let counts: Stream<Burst<u64>> = g.spawn(move |wg| wg.ticker(period).count().limit(n as u32));
+    let counts: Stream<Burst<u64>> = g.spawn(move |wg| wg.ticker(period).count().limit(n));
     let collected = counts.accumulate();
     let mut runner = g.build();
 
@@ -83,10 +83,10 @@ fn spawn_source_delivers_all_values_in_realtime() {
 #[test]
 fn spawn_map_replays_deterministically_in_historical_mode() {
     let period = Duration::from_millis(10);
-    let n = 5u64;
+    let n = 5usize;
 
     let g = GraphBuilder::new();
-    let input = g.ticker(period).count().limit(n as u32); // 1..5 at 0,p,2p,3p,4p
+    let input = g.ticker(period).count().limit(n); // 1..5 at 0,p,2p,3p,4p
     let scaled: Stream<Burst<u64>> =
         input.spawn_map(|s: Stream<Burst<u64>>| s.map(|b: &Burst<u64>| b.iter().sum::<u64>() * 10));
     let collected = scaled
@@ -118,10 +118,10 @@ fn spawn_map_replays_deterministically_in_historical_mode() {
 #[test]
 fn spawn_map_delivers_all_values_in_realtime() {
     let period = Duration::from_millis(2);
-    let n = 5u64;
+    let n = 5usize;
 
     let g = GraphBuilder::new();
-    let input = g.ticker(period).count().limit(n as u32);
+    let input = g.ticker(period).count().limit(n);
     let scaled: Stream<Burst<u64>> =
         input.spawn_map(|s: Stream<Burst<u64>>| s.map(|b: &Burst<u64>| b.iter().sum::<u64>() * 10));
     let collected = scaled.accumulate();
@@ -147,7 +147,7 @@ fn spawn_map_delivers_all_values_in_realtime() {
     // order, grouped arbitrarily. Each output is `sum(group) * 10` over a
     // contiguous run of the input, so walking the expected counts and
     // consuming one run per output must reproduce the input exactly.
-    let mut expected = 1..=n;
+    let mut expected = 1..=n as u64;
     for value in &flat {
         assert_eq!(
             0,
@@ -175,12 +175,12 @@ fn spawn_map_delivers_all_values_in_realtime() {
 #[test]
 fn bounded_spawn_and_spawn_map_match_the_unbounded_result() {
     let period = Duration::from_millis(10);
-    let n = 5u64;
+    let n = 5usize;
 
     let g = GraphBuilder::new();
     // Bound the producer→graph channel to 2 (worker blocks if it gets >2 ahead).
     let counts: Stream<Burst<u64>> =
-        g.spawn_bounded(Some(2), move |wg| wg.ticker(period).count().limit(n as u32));
+        g.spawn_bounded(Some(2), move |wg| wg.ticker(period).count().limit(n));
     let flat = counts.map(|b: &Burst<u64>| b.iter().sum::<u64>());
     // Bound both worker channels to 2 as well.
     let scaled: Stream<Burst<u64>> = flat.spawn_map_bounded(Some(2), |s: Stream<Burst<u64>>| {
